move BatchTextureArray to separate file
