+++ packages/core/src/batch/BatchDrawCall.js
@@ -23,29 +23,3 @@ export class BatchDrawCall
         this.data = null;
     }
 }
-
-/**
- * Used by the batcher to build texture batches.
- * Holds list of textures and their respective locations.
- *
- * @class
- * @memberof PIXI
- */
-export class BatchTextureArray
-{
-    constructor()
-    {
-        this.elements = [];
-        this.ids = [];
-        this.count = 0;
-    }
-
-    clear()
-    {
-        for (let i = 0; i < this.count; i++)
-        {
-            this.elements[i] = null;
-        }
-        this.count = 0;
-    }
-}
