+++ packages/core/src/batch/BatchTextureArray.js
@@ -1,0 +1,37 @@
+/**
+ * Used by the batcher to build texture batches.
+ * Holds list of textures and their respective locations.
+ *
+ * @class
+ * @memberof PIXI
+ */
+export class BatchTextureArray
+{
+    constructor()
+    {
+        /**
+         * inside textures array
+         * @member {PIXI.BaseTexture[]}
+         */
+        this.elements = [];
+        /**
+         * Respective locations for textures
+         * @member {number[]}
+         */
+        this.ids = [];
+        /**
+         * number of filled elements
+         * @member {number}
+         */
+        this.count = 0;
+    }
+
+    clear()
+    {
+        for (let i = 0; i < this.count; i++)
+        {
+            this.elements[i] = null;
+        }
+        this.count = 0;
+    }
+}
